lambda_app/requirements.txt
Pillow==10.1.0
PyPDF2==3.0.1
